refactor(trace-core): restructure packages by role and rename TraceContext to RequestTraceContext
