--- modules/trace/dopamine-trace-core/src/main/kotlin/io/dopamine/trace/core/TraceIdContext.kt
@@ -1,13 +1,0 @@
-package io.dopamine.trace.core
-
-/**
- * Stores and retrieves the current traceId during a request lifecycle.
- *
- * Implementations may use MDC, ThreadLocal, or other context mechanisms
- * depending on the runtime environment.
- */
-interface TraceIdContext {
-    fun getCurrentTraceId(): String?
-    fun setCurrentTraceId(traceId: String)
-    fun clear()
-}
