+++ modules/trace/dopamine-trace-core/src/main/kotlin/io/dopamine/trace/core/generator/TraceIdGenerator.kt
@@ -1,4 +1,4 @@
-package io.dopamine.trace.core
+package io.dopamine.trace.core.generator
 
 /**
  * Generates a unique traceId for request tracking.
