+++ modules/trace/dopamine-trace-core/src/main/kotlin/io/dopamine/trace/core/request/RequestTraceContext.kt
@@ -1,0 +1,13 @@
+package io.dopamine.trace.core.request
+
+/**
+ * Provides access to request-level metadata such as headers or attributes,
+ * used for resolving an incoming traceId.
+ *
+ * Typical implementations may wrap HttpServletRequest or ServerWebExchange.
+ */
+interface RequestTraceContext {
+    fun getHeader(name: String): String?
+
+    fun getAttribute(name: String): String?
+}
