+++ modules/trace/dopamine-trace-core/src/main/kotlin/io/dopamine/trace/core/resolver/CompositeTraceIdResolver.kt
@@ -1,4 +1,6 @@
-package io.dopamine.response.core.trace
+package io.dopamine.trace.core.resolver
+
+import io.dopamine.trace.core.request.RequestTraceContext
 
 /**
  * Composite TraceIdResolver that attempts multiple strategies in order.
@@ -11,7 +13,7 @@ class CompositeTraceIdResolver(
         require(resolvers.isNotEmpty()) { "At least one TraceIdResolver must be provided." }
     }
 
-    override fun resolve(context: TraceContext): String? =
+    override fun resolve(context: RequestTraceContext): String? =
         resolvers
             .asSequence()
             .map { it.resolve(context) }
