+++ modules/trace/dopamine-trace-core/src/main/kotlin/io/dopamine/trace/core/resolver/TraceIdResolver.kt
@@ -1,9 +1,11 @@
-package io.dopamine.response.core.trace
+package io.dopamine.trace.core.resolver
+
+import io.dopamine.trace.core.request.RequestTraceContext
 
 /**
  * Strategy interface for extracting traceId from a request context.
  * - Designed to support multiple environments like Servlet or WebFlux
  */
 fun interface TraceIdResolver {
-    fun resolve(context: TraceContext): String?
+    fun resolve(context: RequestTraceContext): String?
 }
